// src/types/PlayerCard.ts
export interface PlayerCardData {
  id: string | number;
  name: string | number;
  nickname: string | number;
  teamId: string | number;
  teamAbbreviation: string | number;
  playerImage: string;
  gp: string | number;
  w: string | number;
  l: string | number;
  min: string | number;
  pts: string | number;
  ast: string | number;
  reb: string | number;
  fga: string | number;
  fgm: string | number;
  fgPct: string | number;
  fta: string | number;
  ftm: string | number;
  ftPct: string | number;
}